Vim: Pre-populate Git commit message with issue nr
If the branch name contains an issue number at the start, then that can
be used to pre-fill the commit message.

For example, for branches like `ABC-123-make-it-work-better`, it would
pre-populate the commit message with `ABC-123 ` and leave the user in
insert mode.

Maybe there should be a setting to disable this behaviour but eeh.