Vim: Always enable NERDTree
It was a bit annoying that <C-n> starts NERDTree but pressing `-` does
not (mapping for opening NERDTree in the folder of the current file). I
tried to create a Plug mapping for that and use it to enable NERDTree
but it did not work :(. Kept the Plug mapping in though, let's see if it
can be fixed in the future.