Zsh: Add `trash` alias for Linux
The `trash-cli` package can be used to emulate a "trash" bin from the
command-line. Its `trash-put` command can be used to remove things but
still have the possibility to get them back.

On macOS, the `trash` command is installed from Homebrew I think. I use
`trash` in a few scripts so it would be nice to have the same command
available in Linux and macOS.