Vim: Add mapping for disabling lint rule for current line
There is a plugin for this
(https://github.com/chrisgrieser/nvim-rulebook), but it seems a bit
unnecessary to use a plugin for this. I think that my approach is a bit
simpler and doesn't include unneeded things. Anyways, let's see if I
miss any features of the plugin (like viewing the docs).